refactor(skill-pack): unify onboarding flow + strip buildable Ping template
Drop the Track A/B archetype split into a single wallet-as-agent
onboarding flow. Standard shape is `program_id == operator == wallet hex`;
the on-chain `track` enum is purpose-based (Social/Services/Economy/Open),
not implementation-based.

Strip `templates/agent-program-rs/` (a buildable cargo project that
shipped an unmodified Ping service to live testnet on every Track B
follow-through) into `templates/sails-program-layout/` — a non-buildable
annotated layout reference. Real program development now routes to the
companion `vara-skills` pack via `sails-new-app`. Added `compile_error!()`
to the layout reference so accidental `cargo build` fails loudly.

Resume safety: every registration write is preceded by a query
(GetParticipant / GetApplication / ResolveHandle) so re-runs after a
network blip are no-ops. New canonical script
`examples/full_onboarding.sh` bakes in the guards with named exit
constants and an error/rescue table; `smoke.sh --live` invokes it.

Cross-link to the `vara-skills` companion pack from SKILL.md, README,
STARTER_PROMPT, agent-onboarding, the layout reference, and
references/{overview,ownership-model,actor-id-formats}. 12 files now
reference vara-skills (verification minimum was 5).

Lint gates added:
- vara-wallet version detection (E4) with binary-name-anchored regex (C1)
- Track-prose grep gate to prevent vocabulary regression
- No-buildable-templates glob (A1) using `find \( -name X -o -name Y \) -print`
  to dodge POSIX implicit-print-binding silent miss
- LAYOUT REFERENCE ONLY string gate

JSON args switched from heredoc-with-shell-interpolation to `jq -n --arg`
so quoted/backslash chars in user-supplied URLs / handles can't corrupt
the payload before vara-wallet sees it.

Migration note acknowledges that holders of `track: Open` registered
under the old framing have no on-chain patch path (ApplicationPatch
doesn't include `track`), only re-registration under a fresh handle.

Versions bumped 1.0.0 -> 1.1.0 in SKILL.md frontmatter,
marketplace.json metadata.version, and marketplace.json plugins[0].version.

Verified: lint 93/93, smoke offline 8/8.
