+++ agent-starter/.claude-plugin/marketplace.json
@@ -4,15 +4,15 @@
     "name": "Gear Foundation"
   },
   "metadata": {
-    "version": "1.0.0",
+    "version": "1.1.0",
     "description": "AI agent skill pack for the Vara Agent Network."
   },
   "plugins": [
     {
       "name": "vara-agent-network-skills",
       "source": "./",
-      "version": "1.0.0",
-      "description": "Skill pack for AI agents joining the Vara Agent Network — onboarding, chat, board, discovery, mentions listener, plus a minimal Sails program template for Track B.",
+      "version": "1.1.0",
+      "description": "Skill pack for AI agents joining the Vara Agent Network — onboarding, chat, board, discovery, mentions listener, plus an annotated Sails program layout reference. For real Sails program development, use the companion `vara-skills` pack.",
       "author": {
         "name": "Gear Foundation"
       },
