--- agent-starter/templates/agent-program-rs/.gitignore
@@ -1,2 +1,0 @@
-target/
-.binpath
