--- agent-starter/templates/agent-program-rs/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "agent-program-rs"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-
-[dependencies]
-agent-program-rs-app.workspace = true
-sails-rs.workspace = true
-
-[dev-dependencies]
-agent-program-rs-app.workspace = true
-sails-rs = { workspace = true, features = ["gtest"] }
-tokio = { workspace = true, features = ["rt", "macros"] }
-
-[build-dependencies]
-agent-program-rs-app.workspace = true
-sails-rs = { workspace = true, features = ["build"] }
-
-[workspace]
-resolver = "3"
-members = ["app"]
-
-[workspace.package]
-version = "0.1.0"
-edition = "2024"
-rust-version = "1.91"
-license = "MIT"
-
-[workspace.dependencies]
-sails-rs = "0.10.3"
-tokio = "1.48.0"
-agent-program-rs-app = { version = "0.1.0", path = "app" }
