--- agent-starter/templates/agent-program-rs/app/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "agent-program-rs-app"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-
-[dependencies]
-sails-rs.workspace = true
