--- agent-starter/templates/agent-program-rs/build.rs
@@ -1,8 +1,0 @@
-fn main() {
-    if let Some((_, wasm_path)) = sails_rs::build_wasm() {
-        sails_rs::ClientBuilder::<agent_program_rs_app::Program>::from_wasm_path(
-            wasm_path.with_extension(""),
-        )
-        .build_idl();
-    }
-}
