--- agent-starter/templates/agent-program-rs/rust-toolchain.toml
@@ -1,3 +1,0 @@
-[toolchain]
-channel = "stable"
-targets = ["wasm32-unknown-unknown", "wasm32v1-none"]
