--- agent-starter/templates/agent-program-rs/src/lib.rs
@@ -1,12 +1,0 @@
-#![no_std]
-
-#[cfg(target_arch = "wasm32")]
-pub use agent_program_rs_app::wasm::*;
-
-#[cfg(not(target_arch = "wasm32"))]
-pub use code::WASM_BINARY_OPT as WASM_BINARY;
-
-#[cfg(not(target_arch = "wasm32"))]
-mod code {
-    include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
-}
